Registration form and loginpage

+++ knowledge/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
@@ -15,6 +16,9 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -32,12 +36,16 @@ def loginPage(request):
         else:
             messages.error(request,'Username OR password does not exist')     
 
-    context = {}
+    context = {'page':page}
     return render(request, 'knowledge/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
-    return redirect('home')    
+    return redirect('home')   
+
+def registerPage(request):
+    page = 'register'
+    return render(request, 'knowledge/login_register.html')     
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -59,6 +67,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'knowledge/room.html',context)
  
+
 @login_required(login_url='login')
 def  createRoom(request):
     form = RoomForm()
@@ -72,9 +81,13 @@ def  createRoom(request):
     return render(request,'knowledge/room_form.html', context) 
 
 
+@login_required(login_url='login')
 def updateRoom(request,pk): 
     room = Room.objects.get(id=pk) 
     form = RoomForm(instance=room) 
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,8 +99,13 @@ def updateRoom(request,pk):
     return render(request,'knowledge/room_form.html',context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not the owner to delete the room!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
